requirements.txt
altair==5.0.1
ansi2html==1.8.0
appdirs==1.4.4
asttokens==2.2.1
attrs==23.1.0
beautifulsoup4==4.12.2
blinker==1.6.2
cachetools==5.3.1
category-encoders==2.6.1
certifi==2023.5.7
charset-normalizer==3.2.0
click==8.1.5
cloudpickle==2.2.1
cmdstanpy==1.1.0
colorama==0.4.6
comm==0.1.3
contourpy==1.1.0
convertdate==2.4.0
cycler==0.11.0
Cython==0.29.35
dash==2.10.0
debugpy==1.6.7
decorator==5.1.1
Deprecated==1.2.14
deprecation==2.1.0
ephem==4.1.4
feedparser==6.0.10
Flask==2.2.3
fonttools==4.41.0
frozendict==2.3.8
gitdb==4.0.10
GitPython==3.1.32
holidays==0.28
html5lib==1.1
idna==3.4
imbalanced-learn==0.10.1
importlib-metadata==6.8.0
importlib-resources==6.0.0
ipykernel==6.23.1
ipython==8.13.2
ipywidgets==8.0.6
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.18.3
jsonschema-specifications==2023.6.1
jupyter-dash==0.4.2
jupyter_client==8.2.0
jupyter_core==5.3.0
jupyterlab-widgets==3.0.7
kiwisolver==1.4.4
lightgbm==3.3.5
llvmlite==0.40.0
LunarCalendar==0.0.9
lxml==4.9.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.7.2
matplotlib-inline==0.1.6
mdurl==0.1.2
multitasking==0.0.11
nbformat==5.8.0
nest-asyncio==1.5.6
nltk==3.8.1
nsepy==0.8
numba==0.57.0
numpy==1.25.1
packaging==23.1
pandas==1.5.3
pandas-ta==0.3.14b0
patsy==0.5.3
Pillow==9.5.0
plotly==5.15.0
plotly-resampler==0.8.3.2
pmdarima==2.0.3
prophet==1.1.4
protobuf==4.23.4
pyarrow==12.0.1
pycaret==3.0.2
pydeck==0.8.1b0
Pygments==2.15.1
PyMeeus==0.5.12
Pympler==1.0.1
pyod==1.0.9
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
referencing==0.29.1
regex==2023.6.3
requests==2.31.0
rich==13.4.2
rpds-py==0.8.10
scikit-learn==1.2.2
scikit-plot==0.3.7
scipy==1.10.1
sgmllib3k==1.0.0
six==1.16.0
sktime==0.17.0
urllib3==2.0.3
validators==0.20.0
watchdog==3.0.0
webencodings==0.5.1
yellowbrick==1.5
yfinance==0.2.24
zipp==3.16.2
